sexual_content_classifier/requirements.txt
# tensorflow-gpu
tensorflow
pillow
requests
numpy
opencv-python
nudenet
matplotlib
tqdm